Extract Constant from core.py
